fix(core): Check the max number of character before updating its value
This commit tweaks the method called get_max_character_count in the TextImage class and fixes the overflow in images.

The current implementation of the method uses a while loop to increase the maximum number of character that should be rendered per line but the logic inside the loop updates the value of the max_character variable without checking if the new number allow to render the text without overflowing the image.

This commit adds an if statement to check the new value for the max_character variable before updating it.