feat(subscription): Tweak the links to the CL dockets
This commit updates the property called cl_url in the Subscription model. The current implementation of this property computes links that redirect the users to the docket page in CL but We can't add query parameters to this URL because django removes it while redirecting the request, so We're tweaking the property to return the direct link to the docket page and also adding the order_by parameter.