test(TwitterConnector): added test file for TwitterConnector
Added a file for testing TwitterConnector.
These tests might be too specific; they might be too tightly coupled to the internals of TwitterConnector. That is in part a sign that the TwitterConnector internals (methods) can be refactored. It is also in part because I wrote the test to be quite detailed. Depending on the project standards/culture, that level of detail may or may not be appropriate.