test(MastodonConnector): renamed file; + upload_media() and add_status() tests
The 'test_utils.py' was specifically for MastodonConnector. Since there is now also a TwitterConnector, I renamed the file.
I added tests for upload_media() and add_status().
These tests might be too specific; they might be too tightly coupled to the internals of MastodonConnector. That is in part a sign that the MastodonConnector internals (methods) can be refactored. It is also in part because I wrote the test to be quite detailed. Depending on the project standards/culture, that level of detail may or may not be appropriate.